# Reduce string operations memory allocations