--- a/Jint/Native/Array/ArrayExecutionContext.cs
+++ b/Jint/Native/Array/ArrayExecutionContext.cs
@@ -11,15 +11,20 @@ namespace Jint.Native.Array
         // cache key container for array iteration for less allocations
         private static readonly ThreadLocal<ArrayExecutionContext> _executionContext = new ThreadLocal<ArrayExecutionContext>(() => new ArrayExecutionContext());
 
+        private List<uint> _keyCache;
+        private JsValue[] _callArray1;
+        private JsValue[] _callArray3;
+        private JsValue[] _callArray4;
+
         private ArrayExecutionContext()
         {
         }
 
-        public List<uint> KeyCache = new List<uint>();
+        public List<uint> KeyCache => _keyCache = _keyCache ?? new List<uint>();
 
-        public JsValue[] CallArray1 = new JsValue[1];
-        public JsValue[] CallArray3 = new JsValue[3];
-        public JsValue[] CallArray4 = new JsValue[4];
+        public JsValue[] CallArray1 => _callArray1 = _callArray1 ?? new JsValue[1];
+        public JsValue[] CallArray3 => _callArray3 = _callArray3 ?? new JsValue[3];
+        public JsValue[] CallArray4 => _callArray4 = _callArray4 ?? new JsValue[4];
 
         public static ArrayExecutionContext Current => _executionContext.Value;
     }