--- a/Jint/Native/Null.cs
+++ b/Jint/Native/Null.cs
@@ -2,7 +2,7 @@
 {
     public static class Null
     {
-        public readonly static JsValue Instance = JsValue.Null;
-        public readonly static string Text = "null";
+        public static readonly JsValue Instance = JsValue.Null;
+        public const string Text = "null";
     }
 }