… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/TypeConverter.cs