@shestakov 2018-01-03
I would object against this:
> I dropped char and string lookup array to first 128 ASCII characters as beyond that the chars are not.. so common.

In _some_ practical applications _all_ the strings are nonASCII :)